+++ studio/src/app/components/editor/actions/deck/app-actions-deck/app-actions-deck.tsx
@@ -45,6 +45,9 @@ export class AppActionsDeck {
   @Event()
   private selectDeck: EventEmitter<void>;
 
+  @Event()
+  private stepTo: EventEmitter<HTMLElement | undefined>;
+
   private destroyListener;
 
   // Drag and drop is not supported on iOS and Firefox on Android
@@ -72,6 +75,14 @@ export class AppActionsDeck {
       mode: 'ios',
       showBackdrop: false,
       cssClass: 'popover-menu popover-menu-wide'
+    });
+
+    popover.onDidDismiss().then(async ({data}: OverlayEventDetail) => {
+      if (data !== undefined) {
+        const slide: HTMLElement | null = selectDeckSlide(data);
+
+        this.stepTo.emit(slide);
+      }
     });
 
     await popover.present();